Worked-around macOS Ventura beta 7 controller bug.
On beta 7 macOS reports a Xbox Series controller as 2 devices. 1 is the correct product id, and the other is the product id of a Xbox 360 wired controller. We just use the first controller out of the controllers that are connected, so it would sometimes use the 360 one. Now we omit the 360 product id controller, and rumble now works again.