Removed flag that terminates the app when the last window is closed.
The app was shutting down when the user had hid the main window, while a stream window was fullscreen, then closed the stream window. This was unintended behaviour.
Now I've removed that flag and instead have implemented "applicationShouldHandleReopen", so that when the user clicks the app icon in Dock when there is no app window, I now spawn a new main window.