Properly make a new Settings window when closed then opened.
Load settings whenever a settings pane is selected.
Custom resolution, custom fps, and custom artwork dimensions are now optional CGFloats, not non-optional Ints.
We now don't save a resolution or artwork dimension if the width or the height is missing, or 0. You need a valid non-0 number for both for the resolution to be saved.
Custom fps also doesn't accept a missing value or 0.