Switch to a more reliable way of updating the build number.
The previous way, updating the Info.plist directly in the output bundle, using PListBuddy, would sometime fail (race condition?).
This way, the git command to count the commits is run on each build, and a .xcconfig is used to hold the value.
The only downside of this is that this step needs to be run before the build (so that the new value of the .xcconfig file is picked up). The only way that I could see to do that was by adding a Pre-build step to the Scheme. This means that the scheme needs to be Shared, and checked in, and Xcode seems to randomly change the scheme files. (Someone (I can't find who), suggested that always selecting the same target and platform (e.g Any Mac), before committing to git, helps with this.