Move destructive menu action to bottom
This relocates the “Remove host” context menu item to the bottom of the menu and adds a separator before it. Moving the destructive menu action further away from other actions reduces the likelihood that someone will accidentally take a destructive action.

This change is inspired by me accidentally clicking “remove host” instead of “wake PC” just now while trying to connect, and then having to go downstairs and plug a keyboard+monitor in to the PC to re-pair it >.<